update
managed to get the session created and make variables for the queries of tables.... a little confused over what "Create our session (link) from Python to the DB" ... I thought that making the session was doing this, and then closing it shut it down, but I might be wrong. anyway, onward tomorrow!